requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.0
apache-airflow==2.3.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==20.3.0
Babel==2.12.1
blinker==1.6.2
cachelib==0.10.2
cattrs==1.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.2
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.17.3
lazy-object-proxy==1.9.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.2.0
prison==0.2.1
psutil==5.9.5
psycopg==3.0.14
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.9.1
pydantic_core==2.0.1
Pygments==2.15.1
PyJWT==1.7.1
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
requests==2.28.0
requests-toolbelt==1.0.0
rich==13.4.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
unicodecsv==0.14.1
urllib3==1.26.16
Werkzeug==1.0.1
wrapt==1.15.0
WTForms==2.3.3
yarl==1.9.2
